# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/permissions": {
      "post": {
        "operationId": "PermissionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        },
        "tags": [
          "Permissions"
        ]
      },
      "get": {
        "operationId": "PermissionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Permission"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Permissions"
        ]
      }
    },
    "/permissions/{id}": {
      "get": {
        "operationId": "PermissionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        },
        "tags": [
          "Permissions"
        ]
      },
      "patch": {
        "operationId": "PermissionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Permission"
                }
              }
            }
          }
        },
        "tags": [
          "Permissions"
        ]
      },
      "delete": {
        "operationId": "PermissionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Permissions"
        ]
      }
    },
...
    "/messageceo/{id}": {
      "get": {
        "operationId": "MessageCeoController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCeo"
                }
              }
            }
          }
        },
        "tags": [
          "MessageCeo"
        ]
      },
      "patch": {
        "operationId": "MessageCeoController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageCeoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCeo"
                }
              }
            }
          }
        },
        "tags": [
          "MessageCeo"
        ]
      },
      "delete": {
        "operationId": "MessageCeoController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "MessageCeo"
        ]
      }
    },
    "/messageceo/{id}/reply": {
      "patch": {
        "operationId": "MessageCeoController_reply",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageCeo"
                }
              }
            }
          }
        },
        "tags": [
          "MessageCeo"
        ]
      }
    },
...
    "/chat/message/{id}": {
      "delete": {
        "operationId": "ChatController_deleteMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/chat/message/{id}/reply": {
      "post": {
        "operationId": "ChatController_replyToMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/chat/messages/{messageId}": {
      "get": {
        "operationId": "ChatController_getMessage",
        "parameters": [
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/pipeline": {
      "post": {
        "operationId": "PipelineController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineIdea"
                }
              }
            }
          }
        },
        "tags": [
          "Pipeline"
        ]
      },
      "get": {
        "operationId": "PipelineController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departmentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pipeline"
        ]
      }
    },
    "/pipeline/{id}": {
      "get": {
        "operationId": "PipelineController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineIdea"
                }
              }
            }
          }
        },
        "tags": [
          "Pipeline"
        ]
      },
      "patch": {
        "operationId": "PipelineController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineIdea"
                }
              }
            }
          }
        },
        "tags": [
          "Pipeline"
        ]
      },
      "delete": {
        "operationId": "PipelineController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Pipeline"
        ]
      }
    },
    "/pipeline/{id}/react": {
      "post": {
        "operationId": "PipelineController_reactToIdea",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Pipeline"
        ]
      }
    },
    "/pipeline/{id}/comments": {
      "post": {
        "operationId": "PipelineController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineComment"
                }
              }
            }
          }
        },
        "tags": [
          "Pipeline"
        ]
      }
    },
    "/pipeline/comments/{commentId}/react": {
      "post": {
        "operationId": "PipelineController_reactToComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Pipeline"
        ]
      }
    },
...
  },
  "info": {
    "title": "CN Squad API",
    "description": "The CN Squad API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "action",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Staff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hobbies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dateOfBirth": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "Male",
              "Female",
              "Other"
            ],
            "type": "string"
          },
          "maritalStatus": {
            "enum": [
              "Single",
              "Married",
              "Divorced",
              "Widowed"
            ],
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/StaffAddress"
          },
          "employment": {
            "$ref": "#/components/schemas/StaffEmployment"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "resetToken": {
            "type": "string"
          },
          "resetTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "registrationToken": {
            "type": "string"
          },
          "registrationTokenExpires": {
            "format": "date-time",
            "type": "string"
          },
          "isRegistered": {
            "type": "boolean"
          },
          "lastIpAddress": {
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isCeo": {
            "type": "boolean"
          },
          "isOnline": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "uuid",
          "firstName",
          "lastName",
          "description",
          "hobbies",
          "dateOfBirth",
          "gender",
          "email",
          "phone",
          "password",
          "address",
          "employment",
          "roles",
          "isRegistered",
          "isCeo",
          "isOnline",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          }
        },
        "required": [
          "action"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {}
      },
...
      "MessageCeo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sender": {
            "$ref": "#/components/schemas/Staff"
          },
          "replyContent": {
            "type": "string"
          },
          "replyAttachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "repliedAt": {
            "format": "date-time",
            "type": "string"
          },
          "replier": {
            "$ref": "#/components/schemas/Staff"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "subject",
          "content",
          "sender",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateMessageCeoDto": {
        "type": "object",
        "properties": {}
      },
...
      "Attachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "threadReply": {
            "$ref": "#/components/schemas/ThreadReply"
          },
          "url": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "size": {
            "type": "number"
          },
          "mimeType": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "message",
          "threadReply",
          "url",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "text",
              "join",
              "leave",
              "forward"
            ],
            "type": "string"
          },
          "conversation": {
            "$ref": "#/components/schemas/Conversation"
          },
          "author": {
            "$ref": "#/components/schemas/Staff"
          },
          "text": {
            "type": "string"
          },
          "parent": {
            "$ref": "#/components/schemas/Message"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "repliesInThread": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ThreadReply"
            }
          },
          "edited": {
            "type": "boolean"
          },
          "pinned": {
            "type": "boolean"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reaction"
            }
          },
          "mentions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mention"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "repliesInThreadCount": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "type",
          "conversation",
          "author",
          "replies",
          "repliesInThread",
          "edited",
          "pinned",
          "attachments",
          "reactions",
          "mentions",
          "createdAt",
          "updatedAt"
        ]
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "enum": [
              "channel",
              "private",
              "dm",
              "group_dm"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationMember"
            }
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "createdBy": {
            "$ref": "#/components/schemas/Staff"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastMessageAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastMessage": {
            "$ref": "#/components/schemas/Message"
          }
        },
        "required": [
          "id",
          "type",
          "isArchived",
          "members",
          "messages",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PipelineComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "idea": {
            "$ref": "#/components/schemas/PipelineIdea"
          },
          "author": {
            "$ref": "#/components/schemas/Staff"
          },
          "parentComment": {
            "$ref": "#/components/schemas/PipelineComment"
          },
          "content": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineCommentReaction"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "idea",
          "author",
          "parentComment",
          "content",
          "attachments",
          "reactions",
          "createdAt",
          "updatedAt"
        ]
      },
      "PipelineIdea": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pipeRef": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/Staff"
          },
          "department": {
            "$ref": "#/components/schemas/Department"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineReaction"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PipelineComment"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "pipeRef",
          "title",
          "description",
          "author",
          "department",
          "attachments",
          "reactions",
          "comments",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}